requirements.txt
gradio
Pillow
google-generativeai
